Update to 2.0 syntax
The yml format chagnes for dependabot 2.0. This PR updates the output file to match that format.

In addition, update to .NET Core 3.1 (from 3.0).

Finally, the docs repos have a large number of projets (several hundred). There is a 100 limit on the ecosystem nodes in in the file. Add logic to remove projects that don't have dependencies others than the target framework. Dependabot doesn't update those at this time, so removing those keeps the project file small enough for the parser.